+++ bin/main.rb
@@ -1,5 +1,7 @@
 # !/usr/bin/env ruby
-require './lib/game_logic.rb'
+require './lib/game.rb'
+require './lib/player.rb'
+require './lib/board.rb'
 
 puts 'Welcome to the game of Tictoe'
 
@@ -24,32 +26,29 @@ player2 = Player.new('o', 'Player 2')
 
 game = Game.new(board, player1, player2)
 
-
 def make_move(player)
-    loop do
-      puts Rainbow("Enter the #{player.name} move").blue.bright
-      move = gets.chomp.to_i
-      is_valid = player.choose_number(move)
-      unless is_valid
-        puts Rainbow("\nWrong input, try again!\n").red
-        next
-      end
-      break
+  loop do
+    puts Rainbow("Enter the #{player.name} move").blue.bright
+    move = gets.chomp.to_i
+    is_valid = player.choose_number(move)
+    unless is_valid
+      puts Rainbow("\nWrong input, try again!\n").red
+      next
     end
+    break
   end
-
-
-
-def player_turn(player, game, board)
-    make_move(player)
-    puts board.draw_board
-    puts Rainbow(game.check_winner(player)).green.bright
 end
 
+def player_turn(player, game, board)
+  make_move(player)
+  puts board.draw_board
+  puts Rainbow(game.check_winner(player)).green.bright
+end
 
 loop do
-    player_turn(player1,game, board)
-    break if game.winner?(player1.sign)
-    player_turn(player2, game, board)
-    break if game.winner?(player2.sign)
+  player_turn(player1, game, board)
+  break if game.winner?(player1.sign)
+
+  player_turn(player2, game, board)
+  break if game.winner?(player2.sign)
 end
