+++ lib/player.rb
@@ -1,0 +1,28 @@
+class Player
+  attr_accessor :moves
+  attr_reader :sign, :name
+
+  def initialize(sign, name)
+    @sign = sign
+    @moves = []
+    @name = name
+  end
+
+  def choose_number(move)
+    valid = false
+    new_board = Board.board
+    new_board.map! do |number|
+      if number == move
+        @moves.push(number)
+        valid = true
+        @sign
+      else
+        number
+      end
+    end
+
+    Board.change_board(new_board)
+
+    valid
+  end
+end
